// .gadget/server/src/types.ts
export interface ActionExecutionScope {
  recordDeleted?: boolean;
  [key: string]: any;
}

/** Information about the current request being processed by Gadget. */
export interface RequestData {
  /** The remote IP address of the request */
  ip: string;
  /** The requested URL */
  url: string;
  /** The HTTP method for this request (like GET, POST, etc) */
  method: string;
  /** A unique identifier for this request */
  id: string;
  /** The user agent header of the request */
  userAgent?: string;
  /** Headers passed to this request */
  headers: Record<string, string | string[] | undefined>;
}
// ...
export type ActionTrigger =
  | ShopifyWebhookActionTrigger
  | APIActionTrigger
  | ShopifySyncActionTrigger
  | SchedulerActionTrigger
  | BackgroundActionTrigger
  | ShopifyOAuthActionTrigger
  | ShopifyAdminActionTrigger
  | ShopifyCustomerAccountLoginTrigger
  | PlatformTrigger
  | MockActionTrigger
  | GoogleOAuthSignInActionTrigger
  | GoogleOAuthSignUpActionTrigger
  | GadgetEmailPasswordSignUpTrigger
  | GadgetEmailPasswordSignInTrigger
  | GadgetEmailPasswordResetTrigger
  | GadgetEmailPasswordSendResetTrigger
  | GadgetEmailPasswordVerifyTrigger
  | GadgetEmailPasswordSendVerifyTrigger
  | GadgetEmailPasswordChangePasswordTrigger
  | GadgetUserSignOutTrigger
  | BigCommerceOauthActionTrigger
  | BigCommerceUninstallAppTrigger
  | BigCommerceRemoveUserAppTrigger
  | BigCommerceWebhookActionTrigger;

export type ConfigurationVariablesBlob = Record<string, string | null>;

/** @hidden */
export type FindRecordCondition = Record<string, any>;

/**
 * The base attributes most records have
 **/
export interface BaseRecord {
  __typename?: string;
  id?: string;
  state?: any;
  stateHistory?: any;
  [key: string]: any;
}

export interface AnyBulkRecordLoader {
  loadRecord(apiIdentifier: string, namespace: string[], condition: FindRecordCondition): Promise<GadgetRecord<BaseRecord> | undefined>;
}
// ...
export interface NotYetTyped {
  [key: string]: any;
}

export interface AnyParams {
  [key: string]: string | number | boolean | object | bigint | null | undefined;
}

export interface ActionDescriptor {
  key: string;
  type: "Action" | "ModelAction";
  apiIdentifier: string;
  timeoutMilliseconds: number;
  hasReturnType: boolean;
}

export interface GlobalActionDescriptor {
  key: string;
  type: "GlobalAction";
  apiIdentifier: string;
  timeoutMilliseconds: number;
  hasReturnType: boolean;
}

export interface ModelDescriptor extends ModelMetadata {
  validator: {
    validate(context: AnyEffectContext, record: GadgetRecord<BaseRecord>): Promise<void>;
  };
}
/** Represents the data that's in always context */
export interface AnyAmbientContext {
  /** The current request's session, if it has one. Requests made by browsers are given sessions, but requests made using Gadget API Keys are not. */
  session: Session | null;
  /** The current request's session ID, if it has one. Requests made by browsers are given sessions, but requests made using Gadget API Keys are not. */
  sessionID: string | null;
  /** All <%- applicationName %> configuration values */
  config: ConfigurationVariablesBlob;
  /** A map of connection name to instantiated connection objects for <%- applicationName %> */
  connections: Record<string, unknown>;

  /** A signal for if/when the request for processing this unit of work gets prematurely aborted. Useful for passing along to long running requests that should be interrupted when the client goes away. */
  signal: AbortSignal;

  /** A high performance structured logger which writes logs to the Logs Viewer in the Gadget Editor. */
  logger: Logger;
  /**
   * An instance of the API client for <%- applicationName %>.
   *
   * __Note__: This client is authorized using a superuser internal api token and has permission to invoke any action in the system using normal API mutations or the Internal API.
   **/
  api: AnyClient;
  /**
   * The details of the request that is invoking this unit of work, if it was invoked by a request.
   *
   * __Note__: Request details are not always present, like during a background connection sync, a background job, or an action retry.
   **/
  request?: RequestData;
  /**
   * A unique identifier for this context
   */
  id: string;
  /**
   * A boolean describing if this unit of work will be retried automatically or not
   */
  willRetry?: boolean;
  /** App URL for the current environment e.g. https://example.gadget.app */
  currentAppUrl: string;
  /** An object for sending emails */
  emails: GadgetMailer;
  /**
   * The current tenancy for this unit of work
   * @internal
   */
  [AppTenancyKey]?: AppTenancy;
  /** @internal */
  loaders: AnyBulkRecordLoader;
  /** @internal */
  effectAPIs: any;
  /** @internal */
  authConfig?: AuthenticationConfiguration;
}
// ...
export interface BaseActionContext extends AnyAmbientContext {
  /** Details about what triggered this action or global action to run */
  trigger: ActionTrigger;

  /**
   * An object passed between all preconditions and effects of an action execution at the \`scope\` property.
   * Useful for transferring data between effects.
   */
  scope: ActionExecutionScope;

  /**
   * @internal
   */
  [AppTenancyKey]?: AppTenancy;
}

/**
 * An action context type for use in actions that can run on any model.
 */
export interface AnyActionContext extends BaseActionContext {
  type: "action";

  /** Details about the action being executed */
  action: ActionDescriptor;

  /**
   * The record this action is operating on.
   */
  record: GadgetRecord<BaseRecord>;

  /**
   * The model this action is for.
   */
  model: ModelDescriptor;

  /**
   * The parameters passed to the action.
   */
  params: AnyParams;

  /** The phase of execution we're running currently */
  phase: "precondition" | "run" | "success" | "failure";

  /**
   * The current context for this action
   */
  context: AnyActionContext;

  /**
   * @deprecated
   */
  transition?: {
    type: "Transition";
    key: string;
    actionKey: string;
    fromStateKey: string;
    toStateKey: string;
  };
}

/**
 * Describes the context passed to every global action.
 */
export interface AnyGlobalActionContext extends BaseActionContext {
  type: "global-action";

  /** Details about the global action being executed */
  action: GlobalActionDescriptor;

  /** What phase of execution this global action is currently in */
  phase: "precondition" | "run" | "success" | "failure";

  /**
   * The parameters passed to the action.
   */
  params: AnyParams;

  /**
   * The current context for this global action
   */
  context: AnyGlobalActionContext;
}

/**
 * Represents all the context available when executing one effect of an action or global action
 **/
export interface AnyEffectContext extends BaseActionContext {
  type: "effect";
  effect: {
    spec: { id: string };
    configuration: Record<string, any>;
  };
  /**
   * The parameters passed to this effect.
   */
  params: AnyParams;
  /**
   * The specific connection this effect has been contributed by
   **/
  connection?: unknown;
  model?: ModelDescriptor; // present for effects on actions, not present for effects on GlobalActions
  record?: GadgetRecord<BaseRecord>; // present for effects on actions, not present for effects on GlobalActions
}
// ...
export interface FieldMetadata {
  fieldType: string;
  key: string;
  name: string;
  apiIdentifier: string;
  configuration: { [key: string]: any };
  internalWritable: boolean;
}

/**
 * @hidden
 */
export interface ModelMetadata {
  key: string;
  name: string;
  apiIdentifier: string;
  namespace: string[];
  fields: {
    [key: string]: FieldMetadata;
  };
  graphqlTypeName: string;
  stateChart: any;
}

/** Gadget wrapper for NodeMailer, used to facilitate the sending of emails.*/
export interface GadgetMailer {
  /** Verifies SMTP configuration */
  verifyConnection(): Promise<void>;
  /** Sets the transport configuration used for transporting emails */
  setTransport(
    transport: SMTPTransportConfig | SendmailTransportConfig | StreamTransportConfig | JSONTransportConfig | SESTransportConfig
  ): void;
  /** Sends the email using the set transporter */
  sendMail(mailData: MailData): Promise<void>;
  /** Renders an ejs template */
  render(template: string, data: any): string;
}

/**
 * @internal
 */
export interface AuthenticationConfigurationMethod {
  specID: string;
  configuration?:
    | {
        type: "GoogleMethodConfiguration";
        gadgetManagedCredentials: boolean;
        callbackPath: string;
        scopes: string[];
        redirectOnSuccessfulSignInPath?: string | null;
        credentials: {
          clientID?: string;
          clientSecret?: string;
        };
        offlineMode: boolean;
      }
    | {
        type: "EmailPasswordMethodConfiguration";
      };
}

/**
 * @internal
 */
export interface AuthenticationConfiguration {
  signInPath: string;
  redirectOnForbidden: boolean;
  sessionExpirationMs: number;
  defaultAuthRoles: string[];
  redirectOnSuccessfulSignInPath?: string;
  methods?: AuthenticationConfigurationMethod[];
}

/** The options for the mail object being sent */
export type MailData = {
  to: string | string[] | Address | Array<string | Address>;
  subject?: string;
  /** The body of the email */
  html?: string;
  /** The From address displayed to users as to who the email came from. If using Gadget transport, must end in the app's approved subdomain, the value of primaryDomain found in Config. If one is not provided then a default address for the app will be used */
  from?: string | Address;
  /** The Sender header used to identify the agent responsible for the actual transmission of the email. Can only be set when using a Custom transport */
  sender?: string | Address;
};